Fixed a bug in Vector.toUpperCase corrupting nonletters.  Used shared.Random instead of java.util.Random where possible.  Converted RandomReadsMG to Streamer/Writer interface; other minor improvements.  Converted CheckStrand2 to Streamer/Writer.

+++ current/shared/SyncHeart.java
@@ -300,7 +300,7 @@ public class SyncHeart {
 	}
 	
 	/** Thread-safe setter for maximum buffer data */
-	public static void setBufferData(long data) {
+	public static void setBufferData(int data) {
 		writeLock();
 		Shared.setBufferData(data);
 		writeUnlock();
